+++ zed_components/src/zed_camera/src/zed_camera_component_objdet.cpp
@@ -986,7 +986,7 @@ void ZedCamera::stopObjDetect()
     // (e.g clean RVIZ2)
     auto objMsg = std::make_unique<zed_msgs::msg::ObjectsStamped>();
 
-    objMsg->header.stamp = mFrameTimestamp;
+    objMsg->header.stamp = mUsePubTimestamps ? get_clock()->now() : mFrameTimestamp;
     objMsg->header.frame_id = mLeftCamFrameId;
 
     objMsg->objects.clear();
@@ -1106,7 +1106,7 @@ void ZedCamera::processDetectedObjects(rclcpp::Time t)
 
   auto objMsg = std::make_unique<zed_msgs::msg::ObjectsStamped>();
 
-  objMsg->header.stamp = t;
+  objMsg->header.stamp = mUsePubTimestamps ? get_clock()->now() : t;
   objMsg->header.frame_id = mLeftCamFrameId;
 
   objMsg->objects.resize(objCount);